+++ resolwe/flow/models/data.py
@@ -167,6 +167,9 @@ class Data(BaseModel):
     #: error log message
     process_error = ArrayField(models.CharField(max_length=255), default=[])
 
+    #: indicate wether the object was processed by `purge`
+    purged = models.BooleanField(default=False)
+
     #: actual inputs used by the process
     input = JSONField(default=dict)